+++ src/saving_tools.jl
@@ -10,7 +10,8 @@ in `gitpath`, which by default is the currently active project. If the repositor
 is dirty when this function is called the string will end
 with `"_dirty"`.
 
-Return `nothing` if `gitpath` is not a Git repository.
+Return `nothing` if `gitpath` is not a Git repository, i.e. a directory within a git
+repository.
 
 The format of the `git describe` output in general is
 
@@ -41,7 +42,7 @@ julia> gitdescribe(path_to_a_dirty_repo)
 function gitdescribe(gitpath = projectdir())
     # Here we test if the gitpath is a git repository.
     try
-        repo = LibGit2.GitRepo(gitpath)
+        repo = LibGit2.GitRepoExt(gitpath)
     catch er
         @warn "The directory ('$gitpath') is not a Git repository, "*
               "returning `nothing` instead of the commit ID."
@@ -54,7 +55,7 @@ function gitdescribe(gitpath = projectdir())
     end
     # then we return the output of `git describe` or the latest commit hash
     # if no annotated tags are available
-    repo = LibGit2.GitRepo(gitpath)
+    repo = LibGit2.GitRepoExt(gitpath)
     c = try
         gdr = LibGit2.GitDescribeResult(repo)
         fopt = LibGit2.DescribeFormatOptions(dirty_suffix=pointer(suffix))
@@ -70,10 +71,12 @@ end
 
 Generates a patch describing the changes of a dirty repository
 compared to its last commit; i.e. what `git diff HEAD` produces.
+The `gitpath` needs to point to a directory within a git repository,
+otherwise `nothing` is returned.
 """
 function gitpatch(gitpath = projectdir())
     try
-        repo = LibGit2.GitRepo(gitpath)
+        repo = LibGit2.GitRepoExt(gitpath)
     catch er
         @warn "The directory ('$gitpath') is not a Git repository, "*
               "returning `nothing` instead of a patch."
@@ -83,7 +86,7 @@ function gitpatch(gitpath = projectdir())
     # diff = LibGit2.diff_tree(repo, tree)
     # now there is no way to generate the patch with LibGit2.jl.
     # Instead use commands:
-    patch = read(`git --git-dir=$(gitpath)/.git diff HEAD`, String)
+    patch = read(`git -C $(gitpath) diff HEAD`, String)
     return patch
 end